src/Postgres/requirements.txt
mcp
fastapi
uvicorn
psycopg2-binary